Updated variable for organization so that they are consistent between validataions
Fixed bug with checking organization name - gov of canada as it was looking the name with a ";" at the end. But if the department/title was not filled in there would not be a ";". So changed to do the check by checking the first token of a ";" separated string.

Fixed bug with compound messages. i.e. if the gov of canada had errors then all the other validation would also fail. So it could generate 4 error and this was making the errors confusing.  So changes it so that it only generated up to 1 error for each org check.