Disable autofill of layer names in add resources panel
Set the property `loadMapCapabilities` in the online resource panel configuration
so it doesn't try to load the GetCapabilities document of the OGC services and doesn't
show the layers/features grid in the panel.
The user must now fill the Name and Description field by hand.

See https://github.com/geonetwork/core-geonetwork/pull/5763 for more info about this
setting.